Full state managed in AppState
